fixed fabs / abs issue in residual computation for table factors. Removed tab characters from basic reporter and then changed int to size_t in several for loops to address numerous warnings

+++ src/graphlab/metrics/reporters/file_reporter.hpp
@@ -60,7 +60,7 @@ namespace graphlab {
           break;
         case VECTOR:
           fprintf(f, "%s.%s.values=",  ident.c_str(), it->first.c_str());
-          for(int j=0; j<ent.v.size()-1; j++) fprintf(f, "%lf,", ent.v[j]);
+          for(size_t j=0; j<ent.v.size()-1; j++) fprintf(f, "%lf,", ent.v[j]);
           fprintf(f, "%lf\n", ent.v[ent.v.size()-1]);
           fprintf(f, "%s.%s=%lf\n", ident.c_str(), it->first.c_str(),  (ent.value));
           fprintf(f, "%s.%s.count=%d\n", ident.c_str(), it->first.c_str(), ent.count);
